Moved files called by auton chooser back into Auton folder
