+++ src/main/java/frc/robot/commands/auton/CGLaunch1_DriveNone.java
@@ -2,8 +2,11 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.auton;
 import frc.robot.subsystems.Intake;
+import frc.robot.commands.IndexSpeed;
+import frc.robot.commands.IntakeSpeed;
+import frc.robot.commands.LauncherSpeed;
 import frc.robot.subsystems.Index;
 import frc.robot.subsystems.Launcher;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
@@ -12,9 +15,9 @@ import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 /** Autonomous place holder for testing purposes
  * Runs the Launcher for second, then runs the launcher again along with index motors with index, launcher, intake, index
 */ 
-public class AutonLaunch1 extends SequentialCommandGroup {
+public class CGLaunch1_DriveNone extends SequentialCommandGroup {
     
-    public AutonLaunch1(Index indexMotors, Intake intakeMotor, Launcher launcher) {
+    public CGLaunch1_DriveNone(Index indexMotors, Intake intakeMotor, Launcher launcher) {
       addCommands(
       // Start the Launcher - speedFront is first double, speedBack is second
         new LauncherSpeed(launcher, 0.35, 0.40).withTimeout(0.75),
