+++ src/main/java/frc/robot/commands/botchAuton.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.auton;
+package frc.robot.commands;
 import java.util.List;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
 import edu.wpi.first.wpilibj.Timer;
